HFA2/requirements.txt
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.88.0
psycopg2-binary>=2.9.7
python-dateutil>=2.8.2
pytest>=7.4.0 